Refactor Controllers

+++ ChatAppClone/Controllers/UserController.cs
@@ -1,15 +1,14 @@
 ﻿namespace ChatAppClone.Controllers
 {
-    using Microsoft.AspNetCore.Mvc;
-
-    using CloudinaryDotNet.Actions;
-
-    using ChatAppClone.Utilities.Contracts;
     using ChatAppClone.Common.Constants;
+    using ChatAppClone.Common.Messages;
+    using ChatAppClone.Common.Pages;
     using ChatAppClone.Core.Contracts;
     using ChatAppClone.Data.Models;
-    using ChatAppClone.Common.Messages;
     using ChatAppClone.Models.ViewModels.Users;
+    using ChatAppClone.Utilities.Contracts;
+    using CloudinaryDotNet.Actions;
+    using Microsoft.AspNetCore.Mvc;
 
     public class UserController : BaseController
     {
@@ -29,14 +28,21 @@
         {
             try
             {
-                model.Users = await this.userService.GetAsync(this.GetAuthId(), model);
+                var currUserId = this.GetAuth().Item1;
+
+                if (string.IsNullOrWhiteSpace(currUserId))
+                {
+                    return this.RedirectToAction(GeneralPages.Error, GeneralPages.Home, new { statusCode = 401 });
+                }
+
+                model.Users = await this.userService.GetAsync(currUserId, model);
                 model.TotalUsersCount = await this.userService.GetCountAsync(model.SearchTerm);
 
                 return this.View(model);
             }
             catch (Exception)
             {
-                return this.RedirectToAction("Error", "Home", new { area = "", StatusCode = 404 });
+                return this.RedirectToAction(GeneralPages.Error, GeneralPages.Home, new { area = string.Empty, StatusCode = 404 });
             }
         }
 
@@ -55,18 +61,26 @@
 
             try
             {
-                ApplicationUser user = await this.userService.GetByIdAsync(this.GetAuthId());
+                var currUserId = this.GetAuth().Item1;
 
-                ImageUploadResult result = await this.cloudinaryService.UploadPictureAsync(file, CloudinaryConstants.ProfilePicturesFolder, user.ProfilePicturePublicId);
+                if (string.IsNullOrWhiteSpace(currUserId))
+                {
+                    return this.RedirectToAction(GeneralPages.Error, GeneralPages.Home, new { statusCode = 401 });
+                }
 
-                await this.userService.SetProfilePictureAsync(this.GetAuthId(), result.SecureUrl.ToString(), result.PublicId);
+                ApplicationUser user = await this.userService.GetByIdAsync(currUserId);
+
+                ImageUploadResult result = await this.cloudinaryService
+                    .UploadPictureAsync(file, CloudinaryConstants.ProfilePicturesFolder, user.ProfilePicturePublicId);
+
+                await this.userService.SetProfilePictureAsync(currUserId, result.SecureUrl.ToString(), result.PublicId);
+
+                return this.Json(new { success = true });
             }
             catch (Exception ex)
             {
                 return this.Json(new { success = false, error = ex.Message });
             }
-
-            return this.Json(new { success = true });
         }
 
         [HttpPost]
@@ -74,18 +88,25 @@
         {
             try
             {
-                ApplicationUser user = await this.userService.GetByIdAsync(this.GetAuthId());
+                var currUserId = this.GetAuth().Item1;
+
+                if (string.IsNullOrWhiteSpace(currUserId))
+                {
+                    return this.RedirectToAction(GeneralPages.Error, GeneralPages.Home, new { statusCode = 401 });
+                }
+
+                ApplicationUser user = await this.userService.GetByIdAsync(currUserId);
 
                 await this.cloudinaryService.DeletePictureAsync(user.ProfilePicturePublicId!);
 
-                await this.userService.DeleteProfilePictureAsync(this.GetAuthId());
+                await this.userService.DeleteProfilePictureAsync(currUserId);
+
+                return this.Json(new { success = true });
             }
             catch (Exception ex)
             {
                 return this.Json(new { success = false, error = ex.Message });
             }
-
-            return this.Json(new { success = true });
         }
     }
 }
